// src/api/openapi/models/workload.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';


export interface Workload { 
    cost: number | null;
    id: number;
    minutes: number;
    minutes_direction: number;
    user: UserEssential;
}